cosDist cleanup and error fix
cleaned up cosDist function in terms of readability, function names, and code style, and fixed an error in inner production distance calculation with category data weights using the covariate weights

left posing question: covariate information does not use the transpose of y$cvrs, but category information uses the transpose of y$cats.  Surely both are meant to either use or not use the transpose, but which it is is unclear.  If the current state is correct, a clarifying comment should be added to explain why covariate data does not use the transpose and category data does